requirements.txt
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
binance-connector==1.9.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
constantly==15.1.0
cryptography==36.0.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==19.0.0
python-dotenv==0.19.2
requests==2.25.1
service-identity==21.1.0
six==1.16.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.7
zope.interface==5.4.0
